Move the logic for buffering console writes to a separate class
